requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==21.5b1
click==8.0.0
Django==3.2.2
djangorestframework==3.12.4
mypy-extensions==0.4.3
numpy==1.20.3
pandas==1.2.4
pathspec==0.8.1
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
